# New avx512_vnni kernel