--- a/intel_extension_for_transformers/llm/operator/cscr/qbits_ut/weightonly_test.py
+++ b/intel_extension_for_transformers/llm/operator/cscr/qbits_ut/weightonly_test.py
@@ -17,7 +17,7 @@
 import torch
 import inspect
 from functools import wraps
-torch.ops.load_library("../build/libQBits.so")
+torch.ops.load_library("../build/libqbits.so")
 
 
 def capture_args(f):