@luoyu-intel 2023-09-19
q4 group_size=128,  perf on xeon 6248 20cores:

before: 
perf_total_per_op_us[                 MUL_QKV] = 1630.308 ms
perf_total_per_op_us[                FFN_SILU] = 4392.396 ms

after:
perf_total_per_op_us[                 MUL_QKV] = 917.410 ms
perf_total_per_op_us[                FFN_SILU] = 2533.962 ms