@luoyu-intel 2023-09-19
The new kernel introduces an extra reduce buffer(same size as scale). This will slow down the next token inference.
So, the next token generation will use the previous kernel without the reduce buffer.