@luoyu-intel 2023-09-19
q4 group_size=128, perf on xeon 8280 28cores:

before:
perf_total_per_op_us[                 MUL_QKV] = 1176.466 ms
perf_total_per_op_us[                FFN_SILU] = 3134.920 ms

after:
perf_total_per_op_us[                 MUL_QKV] = 659.539 ms
perf_total_per_op_us[                FFN_SILU] = 1788.060 ms